add new script `Set-LocalAdminPassword`
Added a new script I recently created for some team members to quickly update a local admin account used for support purposes. Script can also be helpful for ensuring systems have a standard admin account that can be updated by Windows Laps / Intune when a system is Azure / EntraID Joined or in a on-prem Active Directory environment. Script is documented and has been tested on a virtual machine.